Add support for gathering statistics.
To achieve this, an option to lhttpc:request/{4,5,6,9} and a new
function were introduced.

The new option is {report_stats, Pid, Private} where:
    o  Pid is the pid() of the process who receives gathered statistics
    o  Private is a term() passed along with statistics.

After a call to lhttpc:request/{4,5,6,9} with this option, one can get
the statistics by using the new function lhttpc:get_stats_report/{0,1}.
It must be called from the process Pid specified in the option
parameters. The return values are:
    o  {ok, Statistics, Private}
    o  {error, no_stats}

Statistics is a list of tuples of the form {Step, Data} where:
    o  Step indicates what is measured
    o  Data is a term() containing the statistics

Steps can be:
    o  dns_lookup_time
    o  connection_time
    o  request_send_time
    o  response_wait_time
    o  response_recv_time
    o  total_time
    o  request_headers_length
    o  request_body_length
    o  request_length
    o  response_headers_length
    o  response_body_length
    o  response_length

For *_time, Data has the form {Start, Stop} where:
    o  Start is the start time of the step (erlang:now/0 format)
    o  Stop is the end time

For *_length, Data is an integer indicating the number of bytes.

If the report_stats option is not given, the behaviour is exactly
has before and no overhead should be visible. In this case, the
lhttpc:get_stats_report/{0,1} function will always return {error,
no_stats} (or never returned if the timeout is infinity).

In both cases, the return value of lhttpc:request/{4,5,6,9} doesn't
change.